Revert "Use INT_MAX instead of 65536 as the mesh merge limit."
This reverts commit d69b038b6d18ed092ae2e210cfc6bd601ed094e9.

Causes the atlas to exceed limits.

+++ merge.h
@@ -205,7 +205,7 @@ private:
 	};
 	struct MeshMerge {
 		Vector<MeshState> meshes;
-		int64_t vertex_count = 0;
+		int vertex_count = 0;
 	};
 	static bool setAtlasTexel(void *param, int x, int y, const Vector3 &bar, const Vector3 &, const Vector3 &, float);
 	Ref<Image> dilate(Ref<Image> source_image);
